Added Projectile behavior for player in Josh Scene
Added MousePosition Script that tracked the position of the mouse in world space. This was useful for attaching an object sprite to the mouse when aiming the projectile at a particular location. Added Projectile script which adds functionality relating to the throwing of a projectile with a throwing arc visual to aid the player when aiming. The Player has a limit as to how far they can shoot projectile and can only shoot a projectile every few seconds. Most of the Projectile code was taken from a youtube video with some specific alterations being made for the game. Reference to video can be found in the source code.