Added FadeOut StartScreen and Exit Stats
Added a black fade out at start of game for visual appearance and added an exit point at the end of the level so that when player exits the level, they receive a list of stats based on how well they did. The values and UI presented are placeholders that ideally will be replaced later in development.