Added Actual Enemy bullet actions and other small features
Added Actual Enemy bullet functionality where the actual enemy will shoot a bullet every few seconds at the player until it dies or goes back into reflection. For small features, I added a bag of rocks to the level so that when the player picks it up, they will get the ability to throw projectiles. Added a script called "SetupObjects" that is used for setting up neccessary objects at the start of the game such as disabling projectile behaviour and enabling the FadeOut screen at the start of the level.