Changes to main level and sprites
Painted more of the main level scene. Added all the trees and some extra bits. Re-organized the Tile maps accordingly and with that also fixed the background of the house and other elements. Need to develop the rest of the map. Also tested the map with the player and it works. Added a box collider to "Player with maul" object to stop it from walking over objects